requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
appdirs==1.4.4
astor==0.8.1
asttokens==2.4.1
attrs==23.1.0
backoff==2.2.1
blessed==1.20.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.0
contourpy==1.2.0
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
distro==1.9.0
executing==2.0.1
filelock==3.13.1
fonttools==4.47.0
frozenlist==1.4.1
fsspec==2023.12.2
git-python==1.0.3
gitdb==4.0.11
GitPython==3.1.40
google-api-core==2.15.0
google-api-python-client==2.112.0
google-auth==2.26.1
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.62.0
h11==0.14.0
html2image==2.0.4.3
httpcore==1.0.2
httplib2==0.22.0
httpx==0.26.0
huggingface-hub==0.20.1
idna==3.6
importlib-metadata==7.0.1
inquirer==3.1.4
ipykernel==6.28.0
ipython==8.19.0
jedi==0.19.1
Jinja2==3.1.2
jupyter_client==8.6.0
jupyter_core==5.5.1
kiwisolver==1.4.5
litellm==1.16.7
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
monotonic==1.6
MouseInfo==0.1.3
multidict==6.0.4
nest-asyncio==1.5.8
numpy==1.26.2
oauthlib==3.2.2
open-interpreter==0.2.0
openai==1.6.1
opencv-python==4.9.0.80
outcome==1.3.0.post0
packaging==23.2
parso==0.8.3
pexpect==4.9.0
Pillow==10.1.0
platformdirs==4.1.0
plyer==2.1.0
posthog==3.1.0
prompt-toolkit==3.0.43
protobuf==4.25.1
psutil==5.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyAutoGUI==0.9.54
pydantic==2.5.3
pydantic_core==2.14.6
PyGetWindow==0.0.9
Pygments==2.17.2
PyMonCtl==0.7
PyMsgBox==1.0.9
pyparsing==3.1.1
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
PySocks==1.7.1
pytesseract==0.3.10
python-dateutil==2.8.2
python-dotenv==1.0.0
python-editor==1.0.4
python-xlib==0.33
python3-xlib==0.15
pytweening==1.0.7
PyWinBox==0.6
PyWinCtl==0.3
PyYAML==6.0.1
pyzmq==25.1.2
readchar==4.0.5
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.0
rsa==4.9
selenium==4.16.0
setuptools==69.0.3
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
stack-data==0.6.3
tiktoken==0.4.0
tokenizers==0.15.0
tokentrim==0.1.13
toml==0.10.2
tornado==6.4
tqdm==4.66.1
traitlets==5.14.0
trio==0.23.2
trio-websocket==0.11.1
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==2.1.0
wcwidth==0.2.12
websocket-client==1.7.0
wget==3.2
wsproto==1.2.0
yarl==1.9.4
zipp==3.17.0
